ps3: Window-Based Stereo Matching
(1) Implement the basic stereo algorithm of taking a window
around every pixel in one image, and search for the best match
along the same scan line in the other image using
the sum of squared differences measure.
(2) Using Normalized Correlation measure
(3) Repeat the above 2 for noisy, higher contrast, and a challenge
image.