ps6: Optic Flow: Gaussian and Laplacian Pyramids
Implemented functions to REDUCE, and one that uses it to create
a Gaussian pyramid. Use this to produce a pyramid of 4 levels (0-3),
applying it to the first frame of DataSeq1 sequence.
Here you will also complete the function  create_combined_img(...)
which will output an image that looks like the example below.
- reduce_image(image) -> reduced_image
- gaussian_pyramid(image, levels) -> g_pyr
- create_combined_image(img_list) -> img_out

Signed-off-by: Nikhil Gajendrakumar <nikhil.g@gatech.edu>