ps7: Particle Filter Tracking: Particle Filter
Implemented the particle filter and run it on the pres_debate.mp4 and
noisy_debate.mp4 clip.
I attempted to track Romney’s face. I found the bounding box for his face
for the first frame in template_rect - this is used as template.

Classes, methods:
Defined a class ParticleFilter, with the following methods:
- __init__(frame, template, **kwargs): (constructor) Initialize particle filter object.
- process(frame): Process a frame (image) of video and update filter state.
- render(frame_out): Visualize current particle filter state.

Signed-off-by: Nikhil Gajendrakumar <nikhil.g@gatech.edu>