ps8: Motion History Images: Recognition using MHIs
Computed all the moments for <pq> ∈ {20, 11, 02, 30, 21, 12, 03, 22}.
And, computed this for both the MHI and the thresholded MHI where any
non-zero value is set to 1 (this is the MEI from lecture).
This gives you 16 numbers to describe the MHI.

With this description, did nearest neighbor recognition in a
“leave one out cross validation” recognition test.
Simply tested the MHI & MEI features for each video against the others.
The “nearest” one is the “action” that is recognized.
With the matching method, created a 3x3 “confusion” matrix.

Implemented the Moments class, and compute_feature_difference function

Signed-off-by: Nikhil Gajendrakumar <nikhil.g@gatech.edu>