ps8: Motion History Images: Recognition using MHIs
This time, removed the test subject from the ”training data”.
For example, if we remove subject 1 (P1) from the data,
we will test on all 9 of P1’s sequences and get a confusion matrix for P1.

Then did the same for P2 and P3. Created the confusion matrices for
each of the subjects as described as well as the average confusion matrix.

Signed-off-by: Nikhil Gajendrakumar <nikhil.g@gatech.edu>